move original code to module for testing
